pyproject.toml
[project]
name = "pistachio"
version = "0.1.0"
description = ""
authors = [
    {name = "iamgodot", email = "xugodot@gmail.com"},
]
dependencies = [
    "flask>=2.2.3",
    "flask-sqlalchemy>=3.0.3",
    "gunicorn>=20.1.0",
    "python-dotenv>=1.0.0",
    "gevent>=22.10.2",
    "pyjwt>=2.6.0",
    "requests>=2.29.0",
    "pydantic>=1.10.7",
]
requires-python = ">=3.8"
license = {text = "MIT"}

[tool.pyright]
extraPaths = ["__pypackages__/3.11/lib/"]

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-mock>=3.10.0",
]

[tool.setuptools]
packages = ["pistachio"]

[tool.pytest.ini_options]
log_cli = true
log_level = 10
